+++ R/lamin_cli.R
@@ -352,10 +352,9 @@ lamin_save <- function(filepath, key = NULL, description = NULL, registry = NULL
 
 #' Lamin settings
 #'
-#' Get the current LaminDB settings by calling `lamin settings` on the command
+#' Print the current LaminDB settings by calling `lamin settings` on the command
 #' line
-#'
-#' @returns A list of the current LaminDB settings
+#`
 #' @export
 #'
 #' @examples
@@ -367,61 +366,8 @@ lamin_settings <- function() {
     require_lamindb(silent = TRUE)
     py_config <- reticulate::py_config() # nolint object_usage_linter
 
-    system2("lamin", "settings", stdout = TRUE)
+    system2("lamin", "settings")
   }
 
-  settings_vector <- callr::r(system_fun, show = TRUE, package = "laminr")
-
-  settings <- list()
-  current_parent <- NULL
-
-  for (setting in settings_vector) {
-    # Skip lines that aren't settings
-    if (!grepl(":", setting)) {
-      next
-    }
-
-    is_nested <- grepl("^- ", setting)
-
-    parts <- strsplit(setting, ":", 2)[[1]]
-    field <- trimws(parts[1])
-    value <- trimws(parts[2])
-
-    # Skip empty values
-    if (value == "None" || value == "set()") {
-      next
-    }
-
-    if (value == "True") {
-      value <- TRUE
-    }
-
-    if (value == "False") {
-      value <- FALSE
-    }
-
-    # Turn {'item1', 'item2'} into a vector
-    if (grepl("\\{.*\\}", value)) {
-      value <- gsub(".*\\{(.*?)\\}.*", "\\1", value)
-      value <- strsplit(value, ",")[[1]]
-      value <- trimws(value)
-      value <- gsub("^['\"]|['\"]$", "", value)
-    }
-
-    if (is_nested) {
-      field <- sub("^- ", "", field)
-      settings[[current_parent]][[field]] <- value
-    } else {
-      current_parent <- field
-      settings[[current_parent]] <- list(value = value)
-    }
-  }
-
-  purrr::map(settings, function(.setting) {
-    if (length(.setting) == 1) {
-      .setting[[1]]
-    } else {
-      .setting
-    }
-  })
+  callr::r(system_fun, show = TRUE, package = "laminr") |> invisible()
 }