+++ R/status.R
@@ -19,8 +19,9 @@ laminr_status <- function() {
 
   if (!is.null(default_instance)) {
     # nolint start: object_usage_linter
-    user <- get_current_lamin_user()
-    instance <- suppressMessages(get_current_lamin_instance())
+    settings <- get_current_lamin_settings(minimal = TRUE)
+    user <- settings$user$handle
+    instance <- settings$instance$slug
     # nolint end: object_usage_linter
   }
 
@@ -36,7 +37,10 @@ laminr_status <- function() {
       "wetlab",
       "clinicore",
       "cellregistry",
-      "omop"
+      "omop",
+      "scipy",
+      "numpy",
+      "pandas"
     )
 
     py_versions <- purrr::map_chr(