+++ R/utils.R
@@ -36,6 +36,45 @@ get_default_instance <- function() {
   getOption("LAMINR_DEFAULT_INSTANCE")
 }
 
+#' Get current LaminDB settings
+#'
+#' Get the current LaminDB settings as an R list
+#'
+#' @param minimal If `TRUE`, quickly extract a minimal list of important
+#'  settings instead of converting the complete settings object
+#'
+#' @returns A list of the current LaminDB settings
+#' @export
+#'
+#' @details
+#' This is done using [callr::r()] to avoid importing Python `lamindb` in the
+#' global environment
+get_current_lamin_settings <- function(minimal = FALSE) {
+  call_fun <- function(minimal) {
+    require_lamindb(silent = TRUE)
+    py_ln <- reticulate::import("lamindb")
+
+    py_settings <- py_ln$setup$settings
+
+    if (minimal) {
+      py_builtins <- reticulate::import_builtins()
+      list(
+        instance = list(
+          slug = py_settings$instance$slug,
+          modules = py_builtins$list(py_settings$instance$modules)
+        ),
+        user = list(
+          handle = py_settings$user$handle
+        )
+      )
+    } else {
+      py_settings_to_list(py_settings)
+    }
+  }
+
+  callr::r(call_fun, args = list(minimal = minimal), package = "laminr")
+}
+
 #' Get current LaminDB user
 #'
 #' Get the currently logged in LaminDB user
@@ -45,11 +84,12 @@ get_default_instance <- function() {
 #' @export
 #'
 #' @details
-#' This is done via [lamin_settings()] to avoid importing Python `lamindb`
+#' This is done via [get_current_lamin_settings()] to avoid importing Python
+#' `lamindb`
 get_current_lamin_user <- function() {
-  settings <- lamin_settings()
+  settings <- get_current_lamin_settings(minimal = TRUE)
 
-  handle <- settings[["Current user"]]$handle
+  handle <- settings$user$handle
 
   if (is.null(handle)) {
     cli::cli_alert_danger("No current user")
@@ -68,11 +108,12 @@ get_current_lamin_user <- function() {
 #' @export
 #'
 #' @details
-#' This is done via a [lamin_settings()] to avoid importing Python `lamindb`
+#' This is done via a [get_current_lamin_settings()] to avoid importing Python
+#' `lamindb`
 get_current_lamin_instance <- function() {
-  settings <- lamin_settings()
+  settings <- get_current_lamin_settings(minimal = TRUE)
 
-  instance_slug <- settings[["Current instance"]]$value
+  instance_slug <- settings$instance$slug
 
   if (is.null(instance_slug)) {
     cli::cli_alert_danger("No current instance")